requirements.txt
shiny
folium
geopandas
osmnx
networkx
matplotlib
pandas
shapely
numpy
pandas
pyarrow
requests
scipy
gspread
python-dotenv
